+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/category/Category.java
@@ -149,7 +149,7 @@ public interface Category extends BaseResource, CategoryMixin, com.commercetools
     public String getOrderHint();
 
     /**
-     *  <p>Additional identifier for external systems like Customer Relationship Management (CRM) or Enterprise Resource Planning (ERP).</p>
+     *  <p>Additional identifier for external systems like customer relationship management (CRM) or enterprise resource planning (ERP).</p>
      * @return externalId
      */
 
@@ -297,7 +297,7 @@ public interface Category extends BaseResource, CategoryMixin, com.commercetools
     public void setOrderHint(final String orderHint);
 
     /**
-     *  <p>Additional identifier for external systems like Customer Relationship Management (CRM) or Enterprise Resource Planning (ERP).</p>
+     *  <p>Additional identifier for external systems like customer relationship management (CRM) or enterprise resource planning (ERP).</p>
      * @param externalId value to be set
      */