+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/category/CategoryBuilder.java
@@ -435,7 +435,7 @@ public class CategoryBuilder implements Builder<Category> {
     }
 
     /**
-     *  <p>Additional identifier for external systems like Customer Relationship Management (CRM) or Enterprise Resource Planning (ERP).</p>
+     *  <p>Additional identifier for external systems like customer relationship management (CRM) or enterprise resource planning (ERP).</p>
      * @param externalId value to be set
      * @return Builder
      */
@@ -797,7 +797,7 @@ public class CategoryBuilder implements Builder<Category> {
     }
 
     /**
-     *  <p>Additional identifier for external systems like Customer Relationship Management (CRM) or Enterprise Resource Planning (ERP).</p>
+     *  <p>Additional identifier for external systems like customer relationship management (CRM) or enterprise resource planning (ERP).</p>
      * @return externalId
      */