+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/category/CategoryImpl.java
@@ -205,7 +205,7 @@ public class CategoryImpl implements Category, ModelBase {
     }
 
     /**
-     *  <p>Additional identifier for external systems like Customer Relationship Management (CRM) or Enterprise Resource Planning (ERP).</p>
+     *  <p>Additional identifier for external systems like customer relationship management (CRM) or enterprise resource planning (ERP).</p>
      */
 
     public String getExternalId() {