+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/customer/Customer.java
@@ -85,7 +85,7 @@ public interface Customer extends BaseResource, CustomerMixin, com.commercetools
     public String getCustomerNumber();
 
     /**
-     *  <p>Optional identifier for use in external systems like Customer Relationship Management (CRM) or Enterprise Resource Planning (ERP).</p>
+     *  <p>Optional identifier for use in external systems like customer relationship management (CRM) or enterprise resource planning (ERP).</p>
      * @return externalId
      */
 
@@ -327,7 +327,7 @@ public interface Customer extends BaseResource, CustomerMixin, com.commercetools
     public void setCustomerNumber(final String customerNumber);
 
     /**
-     *  <p>Optional identifier for use in external systems like Customer Relationship Management (CRM) or Enterprise Resource Planning (ERP).</p>
+     *  <p>Optional identifier for use in external systems like customer relationship management (CRM) or enterprise resource planning (ERP).</p>
      * @param externalId value to be set
      */